requirements.txt
apache_beam[gcp]
google-cloud-bigquery
google-cloud-secret-manager
requests
pandas